(FACT-1346) Add default Windows facts as a custom fact
Previously facter doesn't expose default Windows environment variables such as
windir or APPDATA.  This commit adds a custom structured fact which exposes a
whitelisted set of windows environment variables which can be easily consumed in
manifests.  Note that the environment name is purposely uppercased as Windows
env. vars are not case sensitive and it makes it easier to author Puppet
manifests if we're not constantly checking; is TMP or tmp or Tmp.

The custom facts can be accessed via the $facts hash e.g.
$appdata = $facts['windows_env']['APPDATA'].